Improve parsing of non-pedantic emphasis.
The goal of non-pedantic emphasis is to avoid emphasizing
the_middles_of_identifiers. The previous code failed, however, to
do this; for example, in this line

    one_word_with_underscores another_.

emphasis was applied from "underscores" through "another".

Implement a variant of the rules in redcarpet. Redcarpet's rules
say that an underscore matching /\S_[A-Za-z0-9]/ does not count.